# Reduce memory allocations for Process.GetProcessesByName