Fix #40768

Seems felipepessoto's suggestion is not the best choice because there is no any issue on Linux and we have different code paths for local/remote machines on Windows.