@adamsitnik 2020-08-31
> It would be useful to benchmark both time and total allocations for baseline and different variants of the fix.

I agree with @jkotas, we need benchmark results to see how this change is affecting performance

@Serg046 could you please follow the instructions from https://github.com/dotnet/performance/blob/master/docs/benchmarking-workflow-dotnet-runtime.md#Preventing-Regressions and run [these benchmarks](https://github.com/dotnet/performance/blob/master/src/benchmarks/micro/libraries/System.Diagnostics/Perf_Process.cs) against `CoreRun.exe` with and withour your changes?