@Serg046 2020-09-01
@adamsitnik, here we see less allocations but when I ran it next time the number of allocations was almost the same. If we really want to prove that the change is beneficial, we will need a more accurate benchmark with more stable data than the processes and their related code branches. Also please let me know if we want to go deeper in order to further reduce allocations.

Before:
|                     Method |        Job | UnrollFactor |            Mean |           Error |          StdDev |          Median |              Min |             Max |   Gen 0 |   Gen 1 | Gen 2 | Allocated |
|--------------------------- |----------- |------------- |----------------:|----------------:|----------------:|----------------:|-----------------:|----------------:|--------:|--------:|------:|----------:|
|               GetProcesses | Job-WOZLYI |           16 |  2,637,525.2 ns |     9,577.38 ns |     7,997.55 ns |  2,636,506.2 ns |  2,624,027.08 ns |  2,648,383.3 ns | 72.9167 | 20.8333 |     - |  392533 B |
|         GetProcessesByName | Job-WOZLYI |           16 |  2,642,911.2 ns |    13,237.25 ns |    12,382.14 ns |  2,640,937.0 ns |  2,620,994.27 ns |  2,668,576.6 ns | 72.9167 | 20.8333 |     - |  392952 B |

After:
|                     Method |        Job | UnrollFactor |            Mean |           Error |          StdDev |          Median |              Min |             Max |   Gen 0 |   Gen 1 | Gen 2 | Allocated |
|--------------------------- |----------- |------------- |----------------:|----------------:|----------------:|----------------:|-----------------:|----------------:|--------:|--------:|------:|----------:|
|               GetProcesses | Job-SBOBLH |           16 |  2,585,315.5 ns |    14,779.43 ns |    12,341.50 ns |  2,586,216.5 ns |  2,565,340.21 ns |  2,600,253.6 ns | 72.1649 | 30.9278 |     - |  384679 B |
|         GetProcessesByName | Job-SBOBLH |           16 |  2,549,940.4 ns |    36,529.20 ns |    34,169.44 ns |  2,530,560.7 ns |  2,514,150.00 ns |  2,608,208.9 ns | 62.5000 | 17.8571 |     - |  304584 B |

Diff:
| Slower                                                            | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| ----------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Diagnostics.Perf_Activity.EnumerateActivityTagsSmall       |      1.15 |            51.06 |            58.81 |         |
| System.Diagnostics.Perf_Activity.EnumerateActivityTagsLarge       |      1.09 |          8636.64 |          9384.71 |         |
| System.Diagnostics.Perf_Activity.EnumerateActivityTagObjectsLarge |      1.04 |         15011.17 |         15636.22 |         |
| System.Diagnostics.Perf_Activity.EnumerateActivityLinkTagsLarge   |      1.03 |         14204.02 |         14696.13 |         |

| Faster                                                        | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Diagnostics.Perf_Process.GetProcessesByName            |      1.04 |       2640936.98 |       2530560.71 |         |
| System.Diagnostics.Perf_Activity.EnumerateActivityEventsSmall |      1.03 |            96.88 |            94.48 |         |