@danmosemsft 2020-12-15
@adamsitnik should we close this (at least for now)?